Move ComicView
