+++ marinerescue-frontend/src/components/ComicView.js
