+++ marinerescue-frontend/src/components/ComicView.scss
